+++ applications/sample/mapwfs2/index.js
@@ -1,74 +1,7 @@
 /**
- * @class Oskari.paikkatietoikkuna.Main
- *
- * Launcher class for a paikkatietoikkuna.fi map window
- */
-Oskari.clazz.define('Oskari.paikkatietoikkuna.Main', function() {
-
-    this.args = null;
-    this.styleBndl = null;
-}, {
-
-    /**
-     * @method processArgs
-     *
-     * applies page args to this instance
-     */
-    processArgs: function(args) {
-        this.args = args;
-        this.styleBndl = args.style;
-    },
-    /**
-     * @method start
-     *
-     * starts the application with bundle definitions declared
-     * in property appSetup.startupSequence
-     */
-    start: function(cb) {
-
-        var me = this;
-
-        var appSetup = this.appSetup;
-        var appConfig = this.appConfig;
-        var app = Oskari.app;
-
-        app.setApplicationSetup(appSetup);
-        app.setConfiguration(appConfig);
-        app.startApplication(function(startupInfos) {
-            me.instance = startupInfos.bundlesInstanceInfos.mapfull.bundleInstance;
-            if (cb) {
-                cb(me.instance);
-            }
-        });
-    },
-    /**
-     * @static
-     * @property appConfig
-     */
-    appConfig: {
-        // this will be replaced from GetAppSetup
-    },
-
-    /**
-     * @static
-     * @property appSetup.startupSequence
-     */
-    appSetup: {
-
-        // this will be replaced from GetAppSetup
-        startupSequence: []
-    }
-});
-
-/**
  * Start when dom ready
  */
 jQuery(document).ready(function() {
-    var args = {
-        oskariLoaderMode: 'dev',
-        style: 'style1'
-    };
-
     if (!ajaxUrl) {
         alert('Ajax URL not set - cannot proceed');
         return;
@@ -88,7 +21,6 @@ jQuery(document).ready(function() {
 
     // returns empty string if parameter doesn't exist
     // otherwise returns '<param>=<param value>&'
-
     function getAdditionalParam(param) {
         var value = getURLParameter(param);
         if (value) {
@@ -97,30 +29,20 @@ jQuery(document).ready(function() {
         return '';
     }
 
-    var args = {
-        oskariLoaderMode: 'yui',
-        style: 'style1'
-    };
-    if (!ajaxUrl) {
-        alert('Ajax URL not set - cannot proceed');
-        return;
+    // remove host part from url
+    if (ajaxUrl.indexOf('http') == 0) {
+        var hostIdx = ajaxUrl.indexOf('://') + 3;
+        var pathIdx = ajaxUrl.indexOf('/', hostIdx);
+        ajaxUrl = ajaxUrl.substring(pathIdx);
     }
-
+    
     // populate url with possible control parameters
-    ajaxUrl += getAdditionalParam('zoomLevel');
-    ajaxUrl += getAdditionalParam('coord');
-    ajaxUrl += getAdditionalParam('mapLayers');
-    ajaxUrl += getAdditionalParam('oldId');
-    ajaxUrl += getAdditionalParam('viewId');
-
-    ajaxUrl += getAdditionalParam('isCenterMarker');
-    ajaxUrl += getAdditionalParam('address')
-    ajaxUrl += getAdditionalParam('showGetFeatureInfo');
-    ajaxUrl += getAdditionalParam('nationalCadastralReference');
-
-    ajaxUrl += getAdditionalParam('nationalCadastralReferenceHighlight');
-    ajaxUrl += getAdditionalParam('wfsFeature');
-    ajaxUrl += getAdditionalParam('wfsHighlightLayer');
+    var getAppSetupParams = {};
+    if(typeof window.controlParams == 'object') {
+        for(var key in controlParams) {
+            getAppSetupParams[key] = controlParams[key];
+        }
+    }
 
     if (!language) {
         // default to finnish
@@ -128,28 +50,10 @@ jQuery(document).ready(function() {
     }
     Oskari.setLang(language);
 
-
     Oskari.setLoaderMode('dev');
     Oskari.setPreloaded(preloaded);
 
-    // if (location.search && location.search.length > 1) {
-    //     ajaxUrl +=
-    //         location.search.substr(1, location.search.length) + '&';
-    // }
-
-
-    if (args.oskariLoaderAsync && args.oskariLoaderAsync == 'on') {
-        Oskari.setSupportBundleAsync(true);
-    }
-    var main = Oskari.clazz.create('Oskari.paikkatietoikkuna.Main');
-    main.processArgs(args);
-
-    if (ajaxUrl.indexOf('http') == 0) {
-        var hostIdx = ajaxUrl.indexOf('://') + 3;
-        var pathIdx = ajaxUrl.indexOf('/', hostIdx);
-        ajaxUrl = ajaxUrl.substring(pathIdx);
-    }
-    var gfiParamHandler = function(sandbox) {
+    function gfiParamHandler(sandbox) {
         if (getURLParameter('showGetFeatureInfo') != 'true') {
             return;
         }
@@ -162,21 +66,33 @@ jQuery(document).ready(function() {
         });
         sandbox.postRequestByName('MapModulePlugin.GetFeatureInfoRequest', [lon, lat, px.x, px.y]);
     }
+
+    function start(appSetup, appConfig, cb) {
+        var app = Oskari.app;
+
+        app.setApplicationSetup(appSetup);
+        app.setConfiguration(appConfig);
+        app.startApplication(function(startupInfos) {
+            var instance = startupInfos.bundlesInstanceInfos.mapfull.bundleInstance;
+            if (cb) {
+                cb(instance);
+            }
+        });
+
+    }
+
+
     jQuery.ajax({
         type: 'POST',
         dataType: 'json',
-        beforeSend: function(x) {
-            if (x && x.overrideMimeType) {
-                x.overrideMimeType("application/j-son;charset=UTF-8");
-                80
-            }
-        },
-        url: ajaxUrl + 'action_route=GetAppSetup',
-        success: function(appSetup) {
-            if (appSetup.startupSequence && appSetup.configuration) {
-                main.appSetup.startupSequence = appSetup.startupSequence;
-                main.appConfig = appSetup.configuration;
-                main.start(function(instance) {
+        data : getAppSetupParams,
+        url: ajaxUrl + getAdditionalParam('viewId') + 'action_route=GetAppSetup',
+        success: function(app) {
+            if (app.startupSequence && app.configuration) {
+              var appSetup = {
+                "startupSequence": app.startupSequence
+              };
+              start(appSetup, app.configuration, function(instance) {
                     var sb = instance.getSandbox();
                     gfiParamHandler(sb);
                 });