+++ bundles/framework/bundle/mapwfs2/locale/sv.js
@@ -11,6 +11,21 @@ Oskari.registerLocalization({
         "flyout" : {
             "title" : "MapWfs2",
             "message" : "MapWfs2"
+        },
+        "error" : {
+            "title": "Error",
+            "connection_not_available": "Yhteyttä ei voitu muodostaa WFS-taustapalveluun.",
+            "connection_broken": "Yhteys WFS-taustapalveluun katkesi.",
+            "wfs_no_permissions": "WFS-tasoon ({layer}) ei ole riittäviä oikeuksia",
+            "wfs_configuring_layer_failed": "WFS-tason ({layer}) konfigurointi epäonnistui",
+            "wfs_request_failed": "Tason ({layer}) WFS-kysely epäonnistui ",
+            "features_parsing_failed": "WFS-tason ({layer}) featureiden käsittely epäonnistui"
+        },
+        "button": {
+            "cancel": "Avbryt",
+            "show": "Visa",
+            "close": "Stäng",
+            "edit": "Redigera"
         }
     }
 });