+++ bundles/framework/bundle/myplaces2/CategoryHandler.js
@@ -276,6 +276,7 @@ function(instance) {
             name : category.getName(),
             id : category.getId(),
             _isDefault : category.isDefault(),
+            visibleFields: category.getVisibleFields(),
             dot : {
                 size : category.getDotSize(),
                 color : category.getDotColor()
@@ -442,6 +443,9 @@ function(instance) {
         category.setAreaFillColor(values.area.fillColor);
 
         category.setDefault(values._isDefault);
+
+        category.setVisibleFields(values.visibleFields);
+
         return category;
     },
     saveCategory : function(category) {