+++ tools/karma.conf.js
@@ -5,7 +5,6 @@
 // base path, that will be used to resolve files and exclude
 //basePath = '..\..\..\..\..\..\..\..';
 
-
 // list of files / patterns to load in the browser
 files = [
   MOCHA,