+++ ReleaseNotes.md
@@ -10,6 +10,16 @@ A new MapSizeChangedEvent is sent when map size is changed (event is sent on map
 
 Adding base and group layers and sublayers to them is now possible. Also, adding sublayers individually from GetCapabilities query now works.
 
+### statsgrid bundle
+
+- There is now a mode for selecting municipalities from the map instead of the grid. 
+- Checkboxes are visible by default. 
+- Columns can be filtered now by clicking filter link in drop down menu (funnel icon in the header). 
+- Chosen municipalities are now saved to the state.
+
+### layerselector2
+- layer search now supports ontology search. Type min 4 letters and press enter to open up a popup
+
 ## 1.12
 
 ### mapmodule plugin